change type of properties to match what in Header
git-svn-id: http://svn.cern.ch/guest/lhcb/Gauss/trunk@185406 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a9763c123226f8d5a66e3d892947de171b36755f

+++ Gauss/Sim/GaussAlgs/src/GenInit.h
@@ -35,10 +35,10 @@ protected:
 
 private:
   /// Property to set the first event number, if event header is missing
-  int           m_firstEvent;
+  longlong      m_firstEvent;
 
   /// Property to set the run number, if event header is missing
-  int           m_runNumber;
+  unsigned int  m_runNumber;
 
   IGenericTool* m_memoryTool;   ///< Pointer to (private) memory histogram tool
